LayoutTests/media/video-pause-empty-events.html

<video controls></video>
<p>Test that pause() from EMPTY network state triggers load()</p>
<script src=video-test.js></script>
<script>
    video.src = "content/test.mp4";
    testExpected("video.networkState", HTMLMediaElement.EMPTY);
    waitForEventAndEnd('loadstart');
    run("video.pause()");
    consoleWrite("SCRIPT DONE");
</script>
