LayoutTests/media/video-pause-empty-events-expected.txt

Test that pause() from EMPTY network state triggers load()

EXPECTED (video.networkState == '0') OK
RUN(video.pause())
EVENT(loadstart)
END OF TEST
